SPARK-24252: Move Table and TableCapabilities to catalyst module.
This is in preparation for #24246 that adds TableCatalog to catalog.v2
and requires Table in the catalyst module.
