+++ sql/core/src/main/java/org/apache/spark/sql/sources/v2/SupportsRead.java
@@ -17,6 +17,7 @@
 
 package org.apache.spark.sql.sources.v2;
 
+import org.apache.spark.sql.catalog.v2.Table;
 import org.apache.spark.sql.sources.v2.reader.Scan;
 import org.apache.spark.sql.sources.v2.reader.ScanBuilder;
 import org.apache.spark.sql.util.CaseInsensitiveStringMap;
