+++ sql/core/src/main/scala/org/apache/spark/sql/execution/streaming/console.scala
@@ -21,6 +21,7 @@ import java.util
 import java.util.Collections
 
 import org.apache.spark.sql._
+import org.apache.spark.sql.catalog.v2.{Table, TableCapability}
 import org.apache.spark.sql.execution.streaming.sources.ConsoleWrite
 import org.apache.spark.sql.sources.{BaseRelation, CreatableRelationProvider, DataSourceRegister}
 import org.apache.spark.sql.sources.v2._
